requirements.txt
streamlit
langchain
python-dotenv
pydantic
langchain-google-genai
pymupdf